+++ CHANGES.rst
@@ -17,6 +17,7 @@
   - Ruby with ``reek`` [GH-1244]
   - Go with ``megacheck`` [GH-1290]
   - LLVM IR with ``llc`` [GH-1302]
+  - Text prose with ``proselint`` [GH-1304]
 
 - New features: